test: fix the serialization conformance tests gaps (#1172)
* test: add a variable to run only one serialization conformance test


* fix: fixed the encoding of auxiliary data


* test: add a flag to specify the exact re-encoding


* fix: decode metadatum maps as association lists


* test: unit test the decoding of witness sets


* fix: fix the decoding of bytes to accept finite and indefinite length


* fix: fix the decoding of string of definite or indefinite length


* fix: fix the decoding of protocol version


* test: improve the reporting of serialization conformance tests


* test: be more specific than the cddl for the address data


* test: regenerate the serialization fixtures


* fix: fix the decoding of arrays


* fix: fix the decoding of bounded strings


* test: dont twiddle the negative fixtures


* fix: fix the decoding of native scripts


* test: generate correct pool metadata


* test: relabel perturbed test as possibly valid


* fix: fix the decoding of bytes of fixed or max size


* chore: add changelog entry


* chore: rebase on main


* chore: annotate the decoding fixmes for easier retrieval


* fix: transaction index decoding


* test: update the cddl file and regenerate the tests


* chore: rebased on main


* refactoring: vkey to verification_key


* chore: addressed comments


* chore: adding decoding fixtures to git attributes


---------

+++ crates/amaru-kernel/src/cardano/memoized/native_script.rs
@@ -111,7 +111,7 @@ mod tests {
         ScriptPubkey(Hash<KEY>),
         ScriptAll(CborArray<VariableEncodingNativeScript>),
         ScriptAny(CborArray<VariableEncodingNativeScript>),
-        ScriptNOfK(u32, CborArray<VariableEncodingNativeScript>),
+        ScriptNOfK(i64, CborArray<VariableEncodingNativeScript>),
         InvalidBefore(u64),
         InvalidHereafter(u64),
     }
@@ -133,7 +133,7 @@ mod tests {
                 );
 
                 let n_of_k =
-                    (any::<bool>(), any::<u32>(), prop::collection::vec(Self::any(depth - 1), 0..depth as usize))
+                    (any::<bool>(), any::<i64>(), prop::collection::vec(Self::any(depth - 1), 0..depth as usize))
                         .prop_map(|(is_def, n, sigs)| {
                             ScriptNOfK(n, if is_def { CborArray::Def(sigs) } else { CborArray::Indef(sigs) })
                         });
